Enable CONFIG setting for building against external quazip(CMake)
From PR #756, comment was made for need to add commit to CMake Build system files as well.
This commit implements PR#756 into project CMake files.

Executing on the command line:

`cmake ... -DUSE_EXTERN_QUAZIP=ON ...`

Will do the same as PR#756, that is internally set a flag in CMake to test, set a define
for project files during compile, and add the library reference to external libquazip5.

NOTE: Addendum
- altered coverage setting.
Original PR was 'aggressive' in removing thirdparty folder but other files present that are used.