[TvShowModel] Only allow changes to underlying data through model
SeasonModelItem, etc. are not exactly models but only data containers
for the model.

The issue that this commit fixes is with the TvShow model, i.e. the
model that is responsible for displaying all TV shows in a tree-like
widget. When MediaElch reads the list of shows it manages, each show
is appended to the model one at a time. The Qt View/Model for the
TreeView requires us to call `beginInsertRows` before adding a row.
This is required for the TreeView to know that a change occurs and that
the underlying data is about to change. We do this for TV shows only.
When we add seasons or even single episodes to the TV show model, we
used the TvShowModelItem which isn't an QAbstractItemModel but only the
container for the underlying data. The model does not notice that rows
have been added which resulted not only in unsorted TV shows but could
possibly even lead to crashes.
MediaElch added episodes to a TV show *after* the show itself was added
to the view. That worked in the past because none of the shows was
expanded. If however one show was expanded "undefined behavior"
occurred which I wasn't able to understand.
The issue was made more obvious if the feature "Show missing episodes"
was enabled as episodes were added a short period *after* the previous
episodes were added. And that was exactly the first bug in regards to
the TvShowFilesWidget that was found. MediaElch crashed if missing
episodes were loaded and added to the tree.

Long story short: `beginInsertRows` and `endInsertRows` were missing.

This commit changes how MediaElch adds TV shows to the TreeView. All
data changes have to go through the TvShowModel. No other class now has
access to the underlying data (i.e. items like SeasonModelItem).