[DownloadManager] Avoid endless download retries due to race condition
`startNextDownload()` sets `m_downloading` to true. However, when
loading local files it works synchronously and calls
`startNextDownload()` itself again. Because of a guard in the function's
first lines the download fails and never runs successfully. This commit
makes sure that `m_downloading` is then set to `false`.

Furthermore there was a race condition where when re-enqueuing a
download it was possible that another one was already started which
resets m_retries so that we could possibly load files forever. By making
`retries` a member of DownloadManagerElement we can bypass this issue.