[Message] Explicitly set text color to black for message and combo boxes
On Linux distributions that use the KDE Breeze Dark theme (or other dark
themes), the text color of message boxes would be white.  Light text on
light background is not readable. The text is now always set to black.

Furthermore, due to a missing comma (,), two combo boxes were not styled
in light colors and remained dark on the main window.
When a QComboxBox was opened, the background color was also system
dependent. This is now fixed.