[TvTunes] Fix crash when dialog is closed and re-opened for other show
If the TvTunes dialog is opened the first time for a show, it downloads
some resources. This download does not stop when the dialog is closed.
If the dialog is now reopened (e.g. for another show) we get a race
condition because two threads try to dequeue a list with only one item.
Because an empty list cannot be dequeued, `abort()` is called by Qt.

The fix is simple: Use one TvTunes instance for each show download.  We
can achieve that by simply creating a new TV Tunes dialog for each show.