[Export] Fix replacing TV show images in tvshows.html
In a previous commit, images were replaced last for TV shows so that the
inner most block is used for images. This lead to a stupid bug because
the "replaceImages" call was not called unless a `SEASON` block existed.
If it didn't, an early return was triggered.

This commit also introduces some minor changes to image-replacement for
concerts and movies. Image tags are only replaced if the tag is actually
used.

Furthermore, some minor refactoring was done.