[NotificationBox] Hide message and change parent before deleting it
removeWidget() does not change the widget's ownership.
The ownership was transferred to the layout by calling addWidget().
Because we still want to delete the widget, first hide it then
change ownership and only after that delete it.
On systems with Kvantum theme, e.g Manjaro with XFCE, MediaElch
would crash due to SEGFAULT.