[Settings] When canceling the addition of a directory, don't add home
Because we didn't check whether the returned string was empty, the
user's home directory was added per default if the user clicked on
"Cancel" in the file dialog.

On my system this lead to scanning process that took forever.
Especially if the app is sandboxed as it is per default on macOS.