[Project] More changes towards Qt6 compatibility
This commit has some fixes for Qt6.  Some deprecated enums/methods
were replaced.  Comparisons between QStringView and char* required
either QStringView on the right-hand-side or QLatin1String.  We use
the latter now as that's supported on older Qt versions as well.

Some unused `QRegEx` includes were removed.

Some conditionals were used, e.g. for method calls that are required
in Qt5 but are the default (and therefore removed) in Qt6.

This commit does NOT make MediaElch compatible with Qt6.
We still use QRegEx in a lot of places.  Furthermore, we have to wait
for Qt 6.2 to finally get QMultiMediaWidgets.