[UI] Fix large icon column sizes
Some Qt version apparently increased the default "minimum section size"
to 41px. But our icons are just 16px wide so there was a lot of
white-space.  Even though we explicitly set the column width to "22",
it was ignored because the minimum width was higher.

This commit fixes the huge space between icons.