[MovieCrew] Remove temporary "actorsContainer" method
The method was only temporary to make a previous change easier.  We now
replace `actors()` with `actorsContainer()` and remove the latter.

We should work directly on the actors container and not on a vector.

Furthermore, this commit renamed `clear()` to `removeAll()` to
explicitly say that actors are removed and not cleared, i.e. their name
or images re-set.