[MovieFileSearcher] Copy "searcher" data QApplication::processEvents();
It could happen that we run multiple times into `onDirectoryLoaded()`
at once because of the call to `QApplication::processEvents();`.

This also meant that it could happen that the searchers are already
deleted before we return to the normal control flow in the first
call to `onDirectoryLoaded()`.  All references to `searcher` after
`QApplication::processEvents();` were therefore use-after-free bugs.