[MovieFileSearcher] Check if aborted after processEvents() call
After calling `QApplication::processEvents();` we didn't check whether
we were aborted or not.  But it could very well be that the user aborted
the file searcher in the meantime:

```
- onDirectoryLoaded()
  - QApplication::processEvents();
    - User abort()s -> file searchers deleted
  - Function continues: Using invalidated searcher
```

Because we sometimes call `QApplication::processEvents()` in signals as
well, we need to check for `m_aborted` after `emit` statements as well.