[NameFormatter] Use singleton again
There is only one instance of this class (singleton).  While I
mostly dislike singletons, in this case there are reasons.  If we
do not use a singleton and do not store QRegularExpression objects,
each movie that is loaded has to instantiate all QRegularExpressions
for the "exclude word list".  Due to JIT compilation, this has
a _huge_ time disadvantage.  In some tests, 50% of the time went
into the NameFormatter prior to using a singleton.
See https://github.com/Komet/MediaElch/pull/1229 for more details.