[Test] Add new matcher "IsInRange" for signed integers
This matcher can be used where we want to check for an upper and lower
bound.  For example for images. We want to ensure that there are at
least N images but no more than M. If there are more than M, we may
scrape more images than are intended (e.g. non-related images).