[Project] Fix `-Wconversion` warnings for Qt6
This commit introduces new functions `qsizetype_to_int` and
`safe_int_cast` for integer conversion.  In Qt6, the type of `qsizetype`
changes from `int` to `long long int`.  And functions such as `size()`
no longer return `int`.  This resulted in hundreds of conversion
warnings for Qt6.  With this commit, `qsizetype_to_int` does a checked
conversion for Qt6, and simply returns the int in Qt5.

`elch_size_t` is used as `qsizetype` was introduced in Qt 5.10, but we
have to support earlier versions as well.

While most checks are probably not needed (e.g. we don't use more than
2^31 actors), they don't impact performance too much.  Better be safe
than sorry.