[Concert] Fix use-after-free when reloading concerts
There was a use-after-free bug in MediaElch where it would crash if a
concert was selected that has ratings and then "reload" was pressed.

Why did it crash? Because the old concert was deleted, but we still held
the reference to it in the info widget. While clear() was called, we did
not reset the reference.