[UI] Refactor: Remove helper::[set]DevicePixelRatio()
This commit refactors Helper.h: It should not depend on Qt::Widgets.
`helper::devicePixelRatio()` and `helper::setDevicePixelRatio()` were
both not needed.  They were added because of a type mismatch between
`QImage::setDevicePixelRatio()` and `QWidget::devicePixelRatio()`.  The
former expected a qreal whereas the latter returned an int.

Now that we require Qt 5.6, we can use `devicePixelRatioF()` instead.