[Concert] Add `exportTo()` function to ConcertData
By having an `exportTo()` function that is close to the data structure,
we can have a compile-time guarantee that users of ConcertData handle
every field of the structure.

So instead of users pulling data via `concert.title`, they have to
provide an object that has a `exportTitle()` method.  If it does not
exist, it is a compile-time error.  This helps to keep all places
consistent.