refactor: Move different file filters into FileFilter class
The FileFilter class was only used for filtering directories using
QDir's filter functionality.  That's only used as a positive glob
filter.  But we also have file name and directory name filters.  Those
are now also part of it.  Furthermore, rename some methods to make it
more clear what they are, e.g. "filter" is too generic.  Is it
a positive or negative filter? What does it filter?