fix(TV show): Adapt `<episodeguide>` to Kodi v20 changes
Kodi v19.5 and later use a different episodeguide format.  Use that as
well.  Since we don't need to read the episodeguide (since we already
have all unique IDs), we can simply write what we already know: A map
of IDs in JSON format.

Furthermore, remove `year` for Kodi v19 and later, because in v20, only
`premiered` must be used.