refactor(UiUtils): Remove unused button style from `applyStyle()`
`applyStyle()` checked each push button: if it had a `styleType`
property, its style was changed.  But this property was never set
anywhere.  Adding a `qFatal()` to the loop did not result in any
runtime crashes.  Neither did I find the string `styleType` anywhere.

Because the only thing that the function does now is call
`removeFocusRect` if the second argument to `applyStyle()` is true, we
can remove all invocations of this function that don't have a second
argument.