refactor(UiUtils): Remove `applyEffect()`
We added a drop shadow to some buttons if property `dropShadow` is true.
But this was only done on devices with a pixel ratio of 1.  It was never
shown on my current MacBook because its `devicePixelRatio()` is 2.

Even if it is printed, it is slightly visible, but makes MediaElch
inconsistent: We have dozens of buttons that don't have that property.

Therefore, remove the effect.