fix(Movie): Images were not shown (regression)
Follow up to 0491a3cd2f0bc426c0d55dee6372558fdd4cbc35 / #1567:
Due to implicit expectations that we look for certain files in the
movie's file list only if the initial NFO is empty, which is only the
case if we are reloading all movies or there is no NFO in the database.
Otherwise the database takes care to set proper "has image" flags.

That needs to be cleaned up.