test: No longer require explicit resource directory arguments
Via CMake, we now set MediaElch's resource directory in the test
binaries.  That way, we have some default values that are correct most
of the time.  Also create the temporary directory if it not already
exists.  That makes it possible to directly invoke the test binaries
without setup: Make it easier to debug, etc.