fix(UI): Move settings window to top after selecting directory (macOS)
For some reason, on the latest macOS versions (with Qt 5.15.2), the
settings window disappears when a directory was selected.  The main
window is put in front of the settings window.  That is bad, because
we currently do not have a "Windows" bar item where users can easily
switch between windows.