refactor: Remove usage of QML and QtQuick
QtQuick/QML were only used for a single widget: ImageView, which in
turn was only used by the album widget.  It bloated MediaElch and was
one of the places that had trouble with OpenGL, caused GPUs to be used
even if the user didn't want to (macOS), etc.

It duplicated the functionality of ImageGalery.  The only feature
missing is the "cut in half" feature.  Also reordering, but that may
come again at a later point in time.