feat(Music): Use the Discogs API; refactor music scrapers
This commit introduces the Discogs API instead of relying their website.
Furthermore, it refactor the UniversalMusicScraper a bit: Introduce
"scrape and search jobs", similar to how the movie scrapers work.

This process is not finished, but a first start: It removes the burden
of keeping track of each artist/album from the UniversalMusicScraper
and moves it into a scrape/search job per artist/album.