fix(UI): Fix image column count in image dialogs on first appearance
The image dialog window is resized during `exec()`.  That, however,
does not immediately resize the image table.  That happens later, as
children's size is based on the parents size.  This has the effect,
that the column-calculation is based on a lower size than is being shown
to the user.  By postponing the calculation to the next event loop
iteration, we get a proper number of columns.