Fix #1547 Exclusion filter does not exclude subfolder
The problem occurred because QDirIterator was used to generate a single list of files without checking for excluded directories, and filters were only applied to files and their parent directories without checking other directories in the file path. The solution is to use custom traversal. By checking each directory name, it can be skipped if necessary.