fix(movies): Duplicate genres, countries and studio labels when scraping
Tags are set the same way as genres, countries and studios, but didn't have
duplication problem because there was a check if the tag is already set for
the movie. Solution to this problem is to use the same check as used for tags
when adding a new genre, country or studio.