fix(UI): Add "missing" label to "first aired" field if it is the default
For TV shows and episodes, add a marker "missing" to indicate that the
"first aired" date is missing.  Qt does not have a good way to mark the
QDateWidget "invalid".  It requires a date that it can show.

And the defaults are awful: 2000-01-01 is the default.
We simply show the _current_ date with a label "missing" in case the
date is invalid.