feat(HM/ADE): Also load poster back cover
This commits allows users to also load the back cover of movies, not
just the front cover.  Because MediaElch does not support multiple
posters at once, we load the back cover as a backdrop, even though
that is not really correct.  If the user clicks on the poster, the
back cover will also appear.  The same applies to extra fanart.