fix(AppImage): Add workaround for performance bug in Qt
In 516c8a50a84f50c7b8f8565366b9ff2d417fe79d, we changed `onMovieChanged()`
to emit `dataChanged()` for all columns.  That, however, turned out to
be an issue! A performance problem appeared in the AppImage, which uses
Qt5.

According to answers in https://stackoverflow.com/q/53139727/1603627,
Qt uses a quick-path if both "top left" and "bottom right" index are
the same.  If it isn't it may lead to re-rendering of the whole widget.