fix(UI): Fix potential crash when dragging an image onto ImageDialog
For some reason, XCB crashed when running with ASAN.
After debugging a lot, it turns out that using a queued connection
fixes the behavior.  My best guess is that the MyTableWidget instance
was deleted during the execution of of the sigDroppedImage slot, only
to return to MyTableWidget, which was then already deleted.